requirements.txt
ddtrace
fastapi
google-api-core
google-auth
google-cloud-core
google-cloud-secret-manager
pydantic
pydantic-settings
sqlalchemy
uvicorn